file operations
There are pdf files with type and url in an excel file (Excel file is not included here due to data privacy). This is a program that creates folders according to types in the folders named pdf and train in the Excel file, separates and saves the pdf files into folders according to their type, converts each page of the saved files into images and saves them in folders named train according to types, and deletes any empty pages.